refactor: move RSS parsing infrastructure to CelestraKit and fix iOS … (#8)
