+++ .swiftlint.yml
@@ -130,3 +130,9 @@ disabled_rules:
     - opening_brace
     - optional_data_string_conversion
     - pattern_matching_keywords
+custom_rules:
+  no_unchecked_sendable:
+    name: "No Unchecked Sendable"
+    regex: '@unchecked\s+Sendable'
+    message: "Use proper Sendable conformance instead of @unchecked Sendable to maintain strict concurrency safety"
+    severity: error
