+++ Package.resolved
@@ -1,6 +1,24 @@
 {
-  "originHash" : "65f65a345d1475809eb5af8e4db28cb9bc87564b1a60461715c86b10861826bb",
+  "originHash" : "1e288e7fceab587368ce222336eecf4e9bc2be977f835564c7abca2a457363e9",
   "pins" : [
+    {
+      "identity" : "swift-asn1",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/apple/swift-asn1.git",
+      "state" : {
+        "revision" : "810496cf121e525d660cd0ea89a758740476b85f",
+        "version" : "1.5.1"
+      }
+    },
+    {
+      "identity" : "swift-crypto",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/apple/swift-crypto.git",
+      "state" : {
+        "revision" : "95ba0316a9b733e92bb6b071255ff46263bbe7dc",
+        "version" : "3.15.1"
+      }
+    },
     {
       "identity" : "swift-log",
       "kind" : "remoteSourceControl",
@@ -15,17 +33,17 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/brightdigit/SyndiKit.git",
       "state" : {
-        "revision" : "741d082708642a922241bfd75bf06f5264e8d8b9",
-        "version" : "0.7.0"
+        "branch" : "v0.8.0",
+        "revision" : "d324b8abef1f1db86226f4f8556f6bdb7ef90f5b"
       }
     },
     {
       "identity" : "xmlcoder",
       "kind" : "remoteSourceControl",
-      "location" : "https://github.com/brightdigit/XMLCoder",
+      "location" : "https://github.com/CoreOffice/XMLCoder",
       "state" : {
-        "revision" : "8ba70f27664ea8c8b7f38fb4c6f2fd4c129eb9c5",
-        "version" : "1.0.0-alpha.1"
+        "revision" : "5e1ada828d2618ecb79c974e03f79c8f4df90b71",
+        "version" : "0.18.0"
       }
     }
   ],
