+++ Package.swift
@@ -62,8 +62,9 @@ let package = Package(
         ),
     ],
     dependencies: [
-        .package(url: "https://github.com/brightdigit/SyndiKit.git", from: "0.7.0"),
-        .package(url: "https://github.com/apple/swift-log.git", from: "1.0.0")
+        .package(url: "https://github.com/brightdigit/SyndiKit.git", branch: "v0.8.0"),
+        .package(url: "https://github.com/apple/swift-log.git", from: "1.0.0"),
+        .package(url: "https://github.com/apple/swift-crypto.git", from: "3.0.0")
     ],
     targets: [
         // Targets are the basic building blocks of a package, defining a module or a test suite.
@@ -72,7 +73,8 @@ let package = Package(
             name: "CelestraKit",
             dependencies: [
                 .product(name: "SyndiKit", package: "SyndiKit"),
-                .product(name: "Logging", package: "swift-log")
+                .product(name: "Logging", package: "swift-log"),
+                .product(name: "Crypto", package: "swift-crypto")
             ],
             swiftSettings: swiftSettings
         ),
