+++ Sources/CelestraKit/Errors/FeedParserError.swift
@@ -1,5 +1,5 @@
 //
-//  RSSFetcherServiceTests.swift
+//  FeedParserError.swift
 //  CelestraKit
 //
 //  Created by Leo Dion.
@@ -28,13 +28,30 @@
 //
 
 import Foundation
-import Testing
 
-@testable import CelestraKit
+/// Errors that can occur during feed parsing operations
+public enum FeedParserError: Error, Sendable {
+  case invalidURL
+  case invalidArticleURL
+  case missingRequiredData
+  case parsingFailed(underlying: any Error)
+  case networkError(underlying: any Error)
+  case cacheError(underlying: any Error)
 
-#if canImport(FoundationNetworking)
-  import FoundationNetworking
-#endif
-
-/// Namespace for RSSFetcherService tests
-internal enum RSSFetcherServiceTests {}
+  public var localizedDescription: String {
+    switch self {
+    case .invalidURL:
+      return "Invalid or missing feed URL"
+    case .invalidArticleURL:
+      return "Invalid or missing article URL"
+    case .missingRequiredData:
+      return "Required data is missing from the feed"
+    case .parsingFailed(let underlying):
+      return "Feed parsing failed: \(underlying.localizedDescription)"
+    case .networkError(let underlying):
+      return "Network error: \(underlying.localizedDescription)"
+    case .cacheError(let underlying):
+      return "Cache error: \(underlying.localizedDescription)"
+    }
+  }
+}
