+++ Sources/CelestraKit/Models/ParsedFeed.swift
@@ -1,0 +1,85 @@
+//
+//  ParsedFeed.swift
+//  CelestraKit
+//
+//  Created by Leo Dion.
+//  Copyright © 2025 BrightDigit.
+//
+//  Permission is hereby granted, free of charge, to any person
+//  obtaining a copy of this software and associated documentation
+//  files (the "Software"), to deal in the Software without
+//  restriction, including without limitation the rights to use,
+//  copy, modify, merge, publish, distribute, sublicense, and/or
+//  sell copies of the Software, and to permit persons to whom the
+//  Software is furnished to do so, subject to the following
+//  conditions:
+//
+//  The above copyright notice and this permission notice shall be
+//  included in all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+//  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+//  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+//  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+//  HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+//  WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+//  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+//  OTHER DEALINGS IN THE SOFTWARE.
+//
+
+public import Foundation
+
+/// Unified feed model that represents parsed feed data from any format (RSS, Atom, JSON Feed)
+/// This model normalizes data from SyndiKit's Feedable protocol
+public struct ParsedFeed: Identifiable, Sendable, Codable {
+  public let id: URL
+  public let title: String
+  public let subtitle: String?
+  public let url: URL
+  public let siteURL: URL?
+  public let imageURL: URL?
+  public let lastUpdated: Date?
+  public let authors: [ParsedAuthor]
+  public let copyright: String?
+  public let articles: [Article]
+
+  // Specialized properties
+  public let youtubeChannelID: String?
+  public let syndicationUpdate: ParsedSyndicationUpdate?
+
+  // Metadata
+  public let parsedAt: Date
+  public let feedFormat: FeedFormat
+
+  public init(
+    id: URL? = nil,
+    title: String,
+    subtitle: String? = nil,
+    url: URL,
+    siteURL: URL? = nil,
+    imageURL: URL? = nil,
+    lastUpdated: Date? = nil,
+    authors: [ParsedAuthor] = [],
+    copyright: String? = nil,
+    articles: [Article] = [],
+    youtubeChannelID: String? = nil,
+    syndicationUpdate: ParsedSyndicationUpdate? = nil,
+    parsedAt: Date = Date(),
+    feedFormat: FeedFormat = .unknown
+  ) {
+    self.id = id ?? url
+    self.title = title
+    self.subtitle = subtitle
+    self.url = url
+    self.siteURL = siteURL
+    self.imageURL = imageURL
+    self.lastUpdated = lastUpdated
+    self.authors = authors
+    self.copyright = copyright
+    self.articles = articles
+    self.youtubeChannelID = youtubeChannelID
+    self.syndicationUpdate = syndicationUpdate
+    self.parsedAt = parsedAt
+    self.feedFormat = feedFormat
+  }
+}
