+++ Sources/CelestraKit/Models/RobotsRules.swift
@@ -1,5 +1,5 @@
 //
-//  RSSFetcherServiceTests.swift
+//  RobotsRules.swift
 //  CelestraKit
 //
 //  Created by Leo Dion.
@@ -27,14 +27,32 @@
 //  OTHER DEALINGS IN THE SOFTWARE.
 //
 
-import Foundation
-import Testing
+public import Foundation
 
-@testable import CelestraKit
+/// Represents parsed robots.txt rules for a domain
+public struct RobotsRules: Sendable {
+  public let disallowedPaths: [String]
+  public let crawlDelay: TimeInterval?
+  public let fetchedAt: Date
 
-#if canImport(FoundationNetworking)
-  import FoundationNetworking
-#endif
+  /// Check if a given path is allowed
+  public func isAllowed(_ path: String) -> Bool {
+    // If no disallow rules, everything is allowed
+    guard !disallowedPaths.isEmpty else {
+      return true
+    }
 
-/// Namespace for RSSFetcherService tests
-internal enum RSSFetcherServiceTests {}
+    // Check if path matches any disallow rule
+    for disallowedPath in disallowedPaths where path.hasPrefix(disallowedPath) {
+      return false
+    }
+
+    return true
+  }
+
+  public init(disallowedPaths: [String], crawlDelay: TimeInterval?, fetchedAt: Date) {
+    self.disallowedPaths = disallowedPaths
+    self.crawlDelay = crawlDelay
+    self.fetchedAt = fetchedAt
+  }
+}
