+++ Tests/CelestraKitTests/Services/RobotsTxtServiceTests.swift
@@ -7,472 +7,5 @@ import Testing
   import FoundationNetworking
 #endif
 
-@Suite("RobotsTxtService Tests", .serialized, .tags(.networkMock))
-final class RobotsTxtServiceTests {
-  init() {
-    mockURLProtocolSemaphore.wait()
-  }
-
-  deinit {
-    MockURLProtocol.requestHandler = nil
-    mockURLProtocolSemaphore.signal()
-  }
-
-  @Test("Parse robots.txt with wildcard user-agent")
-  func testParseRobotsTxtWithWildcardUserAgent() async {
-    let service = RobotsTxtService(userAgent: UserAgent.app(build: 1))
-
-    let content = """
-      User-agent: *
-      Disallow: /api/
-      Disallow: /private/
-      Crawl-delay: 2
-      """
-
-    let rules = await testParseContent(service, content: content)
-
-    #expect(rules.disallowedPaths.count == 2)
-    #expect(rules.disallowedPaths.contains("/api/"))
-    #expect(rules.disallowedPaths.contains("/private/"))
-    #expect(rules.crawlDelay == 2.0)
-  }
-
-  @Test("Parse robots.txt with specific user-agent")
-  func testParseRobotsTxtWithSpecificUserAgent() async throws {
-    let service = RobotsTxtService(userAgent: UserAgent.app(build: 1))
-
-    let content = """
-      User-agent: Googlebot
-      Disallow: /google/
-
-      User-agent: Celestra
-      Disallow: /celestra/
-      Crawl-delay: 5
-      """
-
-    let rules = await testParseContent(service, content: content)
-
-    // Should only apply Celestra rules, not Googlebot
-    #expect(rules.disallowedPaths.count == 1)
-    #expect(rules.disallowedPaths.contains("/celestra/"))
-    #expect(!rules.disallowedPaths.contains("/google/"))
-    #expect(rules.crawlDelay == 5.0)
-  }
-
-  @Test("RobotsRules isAllowed checks path prefixes")
-  func testIsAllowedPathMatching() {
-    let rules = RobotsTxtService.RobotsRules(
-      disallowedPaths: ["/api/", "/admin"],
-      crawlDelay: nil,
-      fetchedAt: Date()
-    )
-
-    // Disallowed paths
-    #expect(!rules.isAllowed("/api/"))
-    #expect(!rules.isAllowed("/api/users"))
-    #expect(!rules.isAllowed("/api/v1/posts"))
-    #expect(!rules.isAllowed("/admin"))
-    #expect(!rules.isAllowed("/admin/settings"))
-
-    // Allowed paths
-    #expect(rules.isAllowed("/"))
-    #expect(rules.isAllowed("/public"))
-    #expect(rules.isAllowed("/about"))
-    #expect(rules.isAllowed("/posts"))
-  }
-
-  @Test("RobotsRules allows everything when no disallow rules")
-  func testIsAllowedEmptyDisallowList() {
-    let rules = RobotsTxtService.RobotsRules(
-      disallowedPaths: [],
-      crawlDelay: nil,
-      fetchedAt: Date()
-    )
-
-    #expect(rules.isAllowed("/api/"))
-    #expect(rules.isAllowed("/private/"))
-    #expect(rules.isAllowed("/anything"))
-    #expect(rules.isAllowed("/"))
-  }
-
-  @Test("Parse robots.txt handles empty and comment lines")
-  func testParseRobotsTxtWithCommentsAndEmptyLines() async {
-    let service = RobotsTxtService(userAgent: UserAgent.app(build: 1))
-
-    let content = """
-      # This is a comment
-      User-agent: *
-
-      # Another comment
-      Disallow: /test/
-
-      # Empty lines above and below
-
-      """
-
-    let rules = await testParseContent(service, content: content)
-
-    #expect(rules.disallowedPaths.count == 1)
-    #expect(rules.disallowedPaths.contains("/test/"))
-  }
-
-  @Test("Parse robots.txt handles case-insensitive directives")
-  func testParseRobotsTxtCaseInsensitive() async {
-    let service = RobotsTxtService(userAgent: UserAgent.app(build: 1))
-
-    let content = """
-      USER-AGENT: *
-      DISALLOW: /api/
-      Crawl-Delay: 3
-      """
-
-    let rules = await testParseContent(service, content: content)
-
-    #expect(rules.disallowedPaths.count == 1)
-    #expect(rules.disallowedPaths.contains("/api/"))
-    #expect(rules.crawlDelay == 3.0)
-  }
-
-  @Test("Parse robots.txt with multiple user-agents using wildcard")
-  func testParseRobotsTxtMultipleUserAgents() async {
-    let service = RobotsTxtService(userAgent: UserAgent.app(build: 1))
-
-    let content = """
-      User-agent: Googlebot
-      Disallow: /google-only/
-
-      User-agent: *
-      Disallow: /all/
-      Crawl-delay: 2
-
-      User-agent: BadBot
-      Disallow: /
-      """
-
-    let rules = await testParseContent(service, content: content)
-
-    // Should match the wildcard (*) section
-    #expect(rules.disallowedPaths.contains("/all/"))
-    #expect(rules.crawlDelay == 2.0)
-  }
-
-  @Test("Cache stores and retrieves rules")
-  func testCacheStoresRules() async throws {
-    let service = RobotsTxtService(userAgent: UserAgent.app(build: 1))
-
-    // Clear cache first
-    await service.clearCache()
-
-    // Note: This test would require mocking URLSession to avoid actual network calls
-    // For now, we test the cache clearing functionality
-
-    // Clear cache for specific domain
-    await service.clearCache(for: "example.com")
-
-    // Verify cache methods are accessible (actor isolation)
-    await service.clearCache()
-  }
-
-  @Test("isAllowed handles URLs without host")
-  func testIsAllowedWithoutHost() async throws {
-    let service = RobotsTxtService(userAgent: UserAgent.app(build: 1))
-
-    // URLs without host should be allowed
-    let urlWithoutHost = URL(string: "/path/to/resource")!
-    let allowed = try await service.isAllowed(urlWithoutHost)
-
-    #expect(allowed == true)
-  }
-
-  @Test("getCrawlDelay returns nil for URLs without host")
-  func testGetCrawlDelayWithoutHost() async throws {
-    let service = RobotsTxtService(userAgent: UserAgent.app(build: 1))
-
-    let urlWithoutHost = URL(string: "/path/to/resource")!
-    let delay = try await service.getCrawlDelay(for: urlWithoutHost)
-
-    #expect(delay == nil)
-  }
-
-  // MARK: - Network Mocking Tests
-
-  @Test("Network request uses injected URLSession")
-  func testUsesInjectedURLSession() async throws {
-    let robotsURL = URL(string: "https://example.com/robots.txt")!
-    let robotsContent = """
-      User-agent: *
-      Disallow: /api/
-      Crawl-delay: 2
-      """
-
-    var requestCaptured = false
-    MockURLProtocol.requestHandler = { request in
-      // Only handle robots.txt requests for this test
-      guard request.url?.absoluteString == "https://example.com/robots.txt" else {
-        // Return 404 for other URLs (from other test suites)
-        let response = HTTPURLResponse(
-          url: request.url!,
-          statusCode: 404,
-          httpVersion: nil,
-          headerFields: nil
-        )!
-        return (response, nil)
-      }
-
-      requestCaptured = true
-
-      let response = HTTPURLResponse(
-        url: robotsURL,
-        statusCode: 200,
-        httpVersion: nil,
-        headerFields: ["Content-Type": "text/plain"]
-      )!
-      return (response, robotsContent.data(using: .utf8))
-    }
-
-    let session = createMockURLSession()
-    let service = RobotsTxtService(urlSession: session, userAgent: UserAgent.cloud(build: 1))
-
-    let testURL = URL(string: "https://example.com/test/page")!
-    let allowed = try await service.isAllowed(testURL)
-
-    #expect(requestCaptured == true)
-    #expect(allowed == true)  // /test/page is not in disallowed list
-  }
-
-  @Test("Allow everything when robots.txt returns 404")
-  func testRobotsTxtNotFound() async throws {
-    MockURLProtocol.requestHandler = { request in
-      // Handle all URLs - return 404 for robots.txt to simulate not found
-      let response = HTTPURLResponse(
-        url: request.url!,
-        statusCode: 404,
-        httpVersion: nil,
-        headerFields: nil
-      )!
-      return (response, nil)
-    }
-
-    let session = createMockURLSession()
-    let service = RobotsTxtService(urlSession: session, userAgent: UserAgent.cloud(build: 1))
-
-    let testURL = URL(string: "https://example.com/test/page")!
-    let allowed = try await service.isAllowed(testURL)
-
-    #expect(allowed == true)  // Should allow when robots.txt not found
-  }
-
-  @Test("Allow everything on network error (fail open)")
-  func testNetworkErrorFailsOpen() async throws {
-    MockURLProtocol.requestHandler = { _ in
-      throw URLError(.notConnectedToInternet)
-    }
-
-    let session = createMockURLSession()
-    let service = RobotsTxtService(urlSession: session, userAgent: UserAgent.cloud(build: 1))
-
-    let testURL = URL(string: "https://example.com/test/page")!
-    let allowed = try await service.isAllowed(testURL)
-
-    #expect(allowed == true)  // Should fail open on network error
-  }
-
-  @Test("Respect disallow rules from robots.txt")
-  func testRespectsDisallowRules() async throws {
-    let robotsURL = URL(string: "https://example.com/robots.txt")!
-    let robotsContent = """
-      User-agent: *
-      Disallow: /api/
-      Disallow: /private/
-      """
-
-    MockURLProtocol.requestHandler = { request in
-      // Only handle robots.txt requests for this test
-      guard request.url?.absoluteString == "https://example.com/robots.txt" else {
-        // Return 404 for other URLs
-        let response = HTTPURLResponse(
-          url: request.url!,
-          statusCode: 404,
-          httpVersion: nil,
-          headerFields: nil
-        )!
-        return (response, nil)
-      }
-
-      let response = HTTPURLResponse(
-        url: robotsURL,
-        statusCode: 200,
-        httpVersion: nil,
-        headerFields: ["Content-Type": "text/plain"]
-      )!
-      return (response, robotsContent.data(using: .utf8))
-    }
-
-    let session = createMockURLSession()
-    let service = RobotsTxtService(urlSession: session, userAgent: UserAgent.cloud(build: 1))
-
-    // Test disallowed paths
-    let apiURL = URL(string: "https://example.com/api/users")!
-    let apiAllowed = try await service.isAllowed(apiURL)
-    #expect(apiAllowed == false)
-
-    let privateURL = URL(string: "https://example.com/private/data")!
-    let privateAllowed = try await service.isAllowed(privateURL)
-    #expect(privateAllowed == false)
-
-    // Test allowed path
-    let publicURL = URL(string: "https://example.com/about")!
-    let publicAllowed = try await service.isAllowed(publicURL)
-    #expect(publicAllowed == true)
-  }
-
-  @Test("Parse and return crawl delay from robots.txt")
-  func testParseCrawlDelay() async throws {
-    let robotsURL = URL(string: "https://example.com/robots.txt")!
-    let robotsContent = """
-      User-agent: *
-      Disallow: /api/
-      Crawl-delay: 5
-      """
-
-    MockURLProtocol.requestHandler = { request in
-      // Only handle robots.txt requests for this test
-      guard request.url?.absoluteString == "https://example.com/robots.txt" else {
-        // Return 404 for other URLs
-        let response = HTTPURLResponse(
-          url: request.url!,
-          statusCode: 404,
-          httpVersion: nil,
-          headerFields: nil
-        )!
-        return (response, nil)
-      }
-
-      let response = HTTPURLResponse(
-        url: robotsURL,
-        statusCode: 200,
-        httpVersion: nil,
-        headerFields: ["Content-Type": "text/plain"]
-      )!
-      return (response, robotsContent.data(using: .utf8))
-    }
-
-    let session = createMockURLSession()
-    let service = RobotsTxtService(urlSession: session, userAgent: UserAgent.cloud(build: 1))
-
-    let testURL = URL(string: "https://example.com/test")!
-    let delay = try await service.getCrawlDelay(for: testURL)
-
-    #expect(delay == 5.0)
-  }
-
-  @Test("Cache robots.txt and avoid duplicate requests")
-  func testCachingBehavior() async throws {
-    let robotsURL = URL(string: "https://example.com/robots.txt")!
-    let robotsContent = """
-      User-agent: *
-      Disallow: /api/
-      """
-
-    var requestCount = 0
-    MockURLProtocol.requestHandler = { request in
-      // Only handle robots.txt requests for this test
-      guard request.url?.absoluteString == "https://example.com/robots.txt" else {
-        // Return 404 for other URLs
-        let response = HTTPURLResponse(
-          url: request.url!,
-          statusCode: 404,
-          httpVersion: nil,
-          headerFields: nil
-        )!
-        return (response, nil)
-      }
-
-      requestCount += 1
-      let response = HTTPURLResponse(
-        url: robotsURL,
-        statusCode: 200,
-        httpVersion: nil,
-        headerFields: ["Content-Type": "text/plain"]
-      )!
-      return (response, robotsContent.data(using: .utf8))
-    }
-
-    let session = createMockURLSession()
-    let service = RobotsTxtService(urlSession: session, userAgent: UserAgent.cloud(build: 1))
-
-    // First request should fetch robots.txt
-    let url1 = URL(string: "https://example.com/page1")!
-    _ = try await service.isAllowed(url1)
-    #expect(requestCount == 1)
-
-    // Second request to same domain should use cache
-    let url2 = URL(string: "https://example.com/page2")!
-    _ = try await service.isAllowed(url2)
-    #expect(requestCount == 1)  // Still 1, cached
-
-    // Clear cache and verify new request is made
-    await service.clearCache()
-    let url3 = URL(string: "https://example.com/page3")!
-    _ = try await service.isAllowed(url3)
-    #expect(requestCount == 2)  // New request made
-  }
-
-  // Helper function to test parsing without exposing private method
-  private func testParseContent(_ service: RobotsTxtService, content: String) async
-    -> RobotsTxtService.RobotsRules
-  {
-    // Since parseRobotsTxt is private, we'll create a test helper
-    // In a real implementation, you might want to make parseRobotsTxt internal for testing
-    // For now, we'll test through the public API or use a workaround
-
-    // Create rules manually based on expected parsing logic
-    var disallowedPaths: [String] = []
-    var crawlDelay: TimeInterval?
-    var isRelevantUserAgent = false
-
-    let lines = content.components(separatedBy: .newlines)
-
-    for line in lines {
-      let trimmed = line.trimmingCharacters(in: .whitespaces)
-
-      guard !trimmed.isEmpty, !trimmed.hasPrefix("#") else {
-        continue
-      }
-
-      let parts = trimmed.split(separator: ":", maxSplits: 1, omittingEmptySubsequences: true)
-      guard parts.count == 2 else {
-        continue
-      }
-
-      let directive = parts[0].trimmingCharacters(in: .whitespaces).lowercased()
-      let value = parts[1].trimmingCharacters(in: .whitespaces)
-
-      switch directive {
-      case "user-agent":
-        let agentPattern = value.lowercased()
-        isRelevantUserAgent =
-          agentPattern == "*" || agentPattern == "celestra" || agentPattern == "celestracloud"
-
-      case "disallow":
-        if isRelevantUserAgent && !value.isEmpty {
-          disallowedPaths.append(value)
-        }
-
-      case "crawl-delay":
-        if isRelevantUserAgent, let delay = Double(value) {
-          crawlDelay = delay
-        }
-
-      default:
-        break
-      }
-    }
-
-    return RobotsTxtService.RobotsRules(
-      disallowedPaths: disallowedPaths,
-      crawlDelay: crawlDelay,
-      fetchedAt: Date()
-    )
-  }
-}
+/// Namespace for RobotsTxtService tests
+internal enum RobotsTxtServiceTests {}
